Moves cache key logic to manifest and updates references
Centralizes cache key generation within the manifest type to improve encapsulation and simplify cache-related code.
Removes duplicate cache key methods from service logic and updates usages accordingly.
Adds targeted unit tests for cache key generation at the manifest level.
