+++ api/v1beta2/manifest_types_test.go
@@ -1,0 +1,40 @@
+/*
+Copyright 2022.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1beta2_test
+
+import (
+	"testing"
+
+	"github.com/kyma-project/lifecycle-manager/api/shared"
+	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	"github.com/stretchr/testify/require"
+)
+
+func TestGetCacheKey(t *testing.T) {
+	manifest := &v1beta2.Manifest{}
+	manifest.SetName("test-manifest")
+	manifest.SetNamespace("test-namespace")
+	manifest.SetLabels(map[string]string{
+		shared.KymaName: "kyma-test",
+	})
+
+	expectedKey := "kyma-test|test-namespace"
+
+	key, found := manifest.GenerateCacheKey()
+	require.True(t, found)
+	require.Equal(t, expectedKey, key)
+}
